backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "certifi>=2025.10.5",
    "fastapi>=0.119.0",
    "google-auth>=2.41.1",
    "gspread>=6.2.1",
    "httpx>=0.27.0",
    "pydantic>=2.12.3",
    "pymongo>=4.15.3",
    "pytest>=8.0.0",
    "python-dotenv>=1.1.1",
    "supabase>=2.22.0",
    "uvicorn>=0.38.0",
]
